Fix: integer dirichlet value dance from real to int() removed; directly defined as integer

+++ src/test_boundary.f90
@@ -135,7 +135,7 @@ contains
       class(mpi_domain_t), intent(in) :: domain
       integer, dimension(:, :, :), intent(in) :: array
       integer, intent(in) :: bc_types(6)
-      real(kind=sp), intent(in) :: dirichlet_values(6)
+      integer, intent(in) :: dirichlet_values(6)
       logical :: is_bc_face(6)
 
       is_bc_face = domain%is_boundary_face(:)
@@ -144,7 +144,7 @@ contains
       if (is_bc_face(D_WEST)) then
          select case (bc_types(D_WEST))
          case (DIRICHLET)
-            if (any(array(1, 2:size(array, 2) - 1, 2:size(array, 3) - 1) /= int(dirichlet_values(D_WEST)))) then
+            if (any(array(1, 2:size(array, 2) - 1, 2:size(array, 3) - 1) /= dirichlet_values(D_WEST))) then
                call domain%abort("TEST Boundary INT: Not OK: West face Dirichlet")
             end if
          case (NEUMANN)
@@ -161,7 +161,7 @@ contains
       if (is_bc_face(D_EAST)) then
          select case (bc_types(D_EAST))
          case (DIRICHLET)
-            if (any(array(size(array, 1), 2:size(array, 2) - 1, 2:size(array, 3) - 1) /= int(dirichlet_values(D_EAST)))) then
+            if (any(array(size(array, 1), 2:size(array, 2) - 1, 2:size(array, 3) - 1) /= dirichlet_values(D_EAST))) then
                call domain%abort("TEST Boundary INT: Not OK: East face Dirichlet")
             end if
          case (NEUMANN)
@@ -178,7 +178,7 @@ contains
       if (is_bc_face(D_SOUTH)) then
          select case (bc_types(D_SOUTH))
          case (DIRICHLET)
-            if (any(array(2:size(array, 1) - 1, 1, 2:size(array, 3) - 1) /= int(dirichlet_values(D_SOUTH)))) then
+            if (any(array(2:size(array, 1) - 1, 1, 2:size(array, 3) - 1) /= dirichlet_values(D_SOUTH))) then
                call domain%abort("TEST Boundary INT: Not OK: South face Dirichlet")
             end if
          case (NEUMANN)
@@ -195,7 +195,7 @@ contains
       if (is_bc_face(D_NORTH)) then
          select case (bc_types(D_NORTH))
          case (DIRICHLET)
-            if (any(array(2:size(array, 1) - 1, size(array, 2), 2:size(array, 3) - 1) /= int(dirichlet_values(D_NORTH)))) then
+            if (any(array(2:size(array, 1) - 1, size(array, 2), 2:size(array, 3) - 1) /= dirichlet_values(D_NORTH))) then
                call domain%abort("TEST Boundary INT: Not OK: North face Dirichlet")
             end if
          case (NEUMANN)
@@ -212,7 +212,7 @@ contains
       if (is_bc_face(D_LOW)) then
          select case (bc_types(D_LOW))
          case (DIRICHLET)
-            if (any(array(2:size(array, 1) - 1, 2:size(array, 2) - 1, 1) /= int(dirichlet_values(D_LOW)))) then
+            if (any(array(2:size(array, 1) - 1, 2:size(array, 2) - 1, 1) /= dirichlet_values(D_LOW))) then
                call domain%abort("TEST Boundary INT: Not OK: Low face Dirichlet")
             end if
          case (NEUMANN)
@@ -229,7 +229,7 @@ contains
       if (is_bc_face(D_HIGH)) then
          select case (bc_types(D_HIGH))
          case (DIRICHLET)
-            if (any(array(2:size(array, 1) - 1, 2:size(array, 2) - 1, size(array, 3)) /= int(dirichlet_values(D_HIGH)))) then
+            if (any(array(2:size(array, 1) - 1, 2:size(array, 2) - 1, size(array, 3)) /= dirichlet_values(D_HIGH))) then
                call domain%abort("TEST Boundary INT: Not OK: High face Dirichlet")
             end if
          case (NEUMANN)
